requirements.txt
Django==1.9.4
telepot==6.6
gunicorn==19.4.5
djangorestframework==3.3.3
structlog==16.0.0